+++ xmonad-manage/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## master
+- Added `remove` command, which removes specified profile.
 
 ## 0.4
 (Changelog from this point is lost)